Revert "Change offset type from Int64 to Int in Lyrics struct for JellyfinAPI compatibility"
This reverts commit b6c39bb0efc2a6ce6ea18bf56e3d7db495f0efea.

+++ Sources/EtoileKit/Data/Lyrics.swift
@@ -10,6 +10,6 @@ import JellyfinAPI
 
 public struct Lyrics {
     public let lines: [LyricLine]?
-    public let offset: Int? // Has to be Int because JellyfinAPI package doesn't support Int64
+    public let offset: Int64?
     public let synced: Bool?
 }
